AVR: fix get end stop state.

+++ hal/microchip/avr/rumba/board_cfg.h
@@ -112,7 +112,7 @@
 #define CPU_MS_TIMER_IRQ_1            0x02
 #define CPU_MS_TIMER_IRQ_0            0
 #define CPU_MS_TIMER_NEEDS_CNT        8
-#define CPU_MS_TIMER_RELOAD_VALUE     (F_CPU/1000)/8
+#define CPU_MS_TIMER_RELOAD_VALUE     (uint8_t)((F_CPU/1000)/8)
 #define CPU_MS_TIMER_COMPARE_ISR      TIMER0_COMPA_vect
 
 // PWM
